LPS-80060 Move to the right package
