+++ modules/apps/forms-and-workflow/dynamic-data-lists/dynamic-data-lists-web/src/main/java/com/liferay/dynamic/data/lists/web/internal/template/DDLDisplayTemplateHandler.java
@@ -22,8 +22,8 @@ import com.liferay.dynamic.data.lists.service.DDLRecordLocalService;
 import com.liferay.dynamic.data.lists.service.DDLRecordService;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetLocalService;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetService;
-import com.liferay.dynamic.data.lists.web.configuration.DDLWebConfigurationKeys;
-import com.liferay.dynamic.data.lists.web.configuration.DDLWebConfigurationUtil;
+import com.liferay.dynamic.data.lists.web.internal.configuration.DDLWebConfigurationKeys;
+import com.liferay.dynamic.data.lists.web.internal.configuration.DDLWebConfigurationUtil;
 import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
 import com.liferay.dynamic.data.mapping.service.DDMStructureService;
 import com.liferay.dynamic.data.mapping.service.DDMTemplateLocalService;
